Split H264 request building into user codecs
