+++ user/lib/Makefile.toml
@@ -7,11 +7,11 @@ skip_core_tasks = true
 
 [tasks.fmt]
 description = "Format all user library code"
-dependencies = ["fmt-scarlet-abi", "fmt-scarlet-sys", "fmt-scarlet-rt", "fmt-scarlet-os", "fmt-std", "fmt-framebuffer", "fmt-gpu", "fmt-sws-protocol", "fmt-scarlet-desktop-config", "fmt-scpm"]
+dependencies = ["fmt-scarlet-abi", "fmt-scarlet-sys", "fmt-scarlet-rt", "fmt-scarlet-os", "fmt-scarlet-codecs", "fmt-std", "fmt-framebuffer", "fmt-gpu", "fmt-sws-protocol", "fmt-scarlet-desktop-config", "fmt-scpm"]
 
 [tasks.fmt-check]
 description = "Check formatting for all user library code"
-dependencies = ["fmt-scarlet-abi-check", "fmt-scarlet-sys-check", "fmt-scarlet-rt-check", "fmt-scarlet-os-check", "fmt-std-check", "fmt-framebuffer-check", "fmt-gpu-check", "fmt-sws-protocol-check", "fmt-scpm-check"]
+dependencies = ["fmt-scarlet-abi-check", "fmt-scarlet-sys-check", "fmt-scarlet-rt-check", "fmt-scarlet-os-check", "fmt-scarlet-codecs-check", "fmt-std-check", "fmt-framebuffer-check", "fmt-gpu-check", "fmt-sws-protocol-check", "fmt-scpm-check"]
 
 [tasks.fmt-scarlet-abi-check]
 description = "Check formatting for Scarlet ABI code"
@@ -34,6 +34,12 @@ args = ["fmt", "--", "--check"]
 [tasks.fmt-scarlet-os-check]
 description = "Check formatting for Scarlet OS wrapper code"
 cwd = "scarlet-os"
+command = "cargo"
+args = ["fmt", "--", "--check"]
+
+[tasks.fmt-scarlet-codecs-check]
+description = "Check formatting for Scarlet codec request builder code"
+cwd = "scarlet-codecs"
 command = "cargo"
 args = ["fmt", "--", "--check"]
 
@@ -91,6 +97,12 @@ cwd = "scarlet-os"
 command = "cargo"
 args = ["fmt"]
 
+[tasks.fmt-scarlet-codecs]
+description = "Format Scarlet codec request builder code"
+cwd = "scarlet-codecs"
+command = "cargo"
+args = ["fmt"]
+
 [tasks.fmt-framebuffer]
 description = "Format user library framebuffer code"
 cwd = "framebuffer"
@@ -138,6 +150,7 @@ script = [
     "cd ../scarlet-sys && cargo build --target ../../targets/riscv64gc-unknown-scarlet-elf.json",
     "cd ../scarlet-rt && cargo build --target ../../targets/riscv64gc-unknown-scarlet-elf.json",
     "cd ../scarlet-os && cargo build --target ../../targets/riscv64gc-unknown-scarlet-elf.json",
+    "cd ../scarlet-codecs && cargo build --target ../../targets/riscv64gc-unknown-scarlet-elf.json",
     "cd ../std && cargo build --target ../../targets/riscv64gc-unknown-scarlet-elf.json",
     "cd ../framebuffer && cargo build --target ../../targets/riscv64gc-unknown-scarlet-elf.json",
     "cd ../gpu && cargo build --target ../../targets/riscv64gc-unknown-scarlet-elf.json",
@@ -153,6 +166,7 @@ script = [
     "cd ../scarlet-sys && cargo build --release --target ../../targets/riscv64gc-unknown-scarlet-elf.json",
     "cd ../scarlet-rt && cargo build --release --target ../../targets/riscv64gc-unknown-scarlet-elf.json",
     "cd ../scarlet-os && cargo build --release --target ../../targets/riscv64gc-unknown-scarlet-elf.json",
+    "cd ../scarlet-codecs && cargo build --release --target ../../targets/riscv64gc-unknown-scarlet-elf.json",
     "cd ../std && cargo build --release --target ../../targets/riscv64gc-unknown-scarlet-elf.json",
     "cd ../framebuffer && cargo build --release --target ../../targets/riscv64gc-unknown-scarlet-elf.json",
     "cd ../gpu && cargo build --release --target ../../targets/riscv64gc-unknown-scarlet-elf.json",
@@ -168,6 +182,7 @@ script = [
     "cd ../scarlet-sys && cargo build --target ../../targets/aarch64-unknown-scarlet-elf.json",
     "cd ../scarlet-rt && cargo build --target ../../targets/aarch64-unknown-scarlet-elf.json",
     "cd ../scarlet-os && cargo build --target ../../targets/aarch64-unknown-scarlet-elf.json",
+    "cd ../scarlet-codecs && cargo build --target ../../targets/aarch64-unknown-scarlet-elf.json",
     "cd ../std && cargo build --target ../../targets/aarch64-unknown-scarlet-elf.json",
     "cd ../framebuffer && cargo build --target ../../targets/aarch64-unknown-scarlet-elf.json",
     "cd ../gpu && cargo build --target ../../targets/aarch64-unknown-scarlet-elf.json",
@@ -183,6 +198,7 @@ script = [
     "cd ../scarlet-sys && cargo build --release --target ../../targets/aarch64-unknown-scarlet-elf.json",
     "cd ../scarlet-rt && cargo build --release --target ../../targets/aarch64-unknown-scarlet-elf.json",
     "cd ../scarlet-os && cargo build --release --target ../../targets/aarch64-unknown-scarlet-elf.json",
+    "cd ../scarlet-codecs && cargo build --release --target ../../targets/aarch64-unknown-scarlet-elf.json",
     "cd ../std && cargo build --release --target ../../targets/aarch64-unknown-scarlet-elf.json",
     "cd ../framebuffer && cargo build --release --target ../../targets/aarch64-unknown-scarlet-elf.json",
     "cd ../gpu && cargo build --release --target ../../targets/aarch64-unknown-scarlet-elf.json",
@@ -198,6 +214,7 @@ script = [
     "cd ../scarlet-sys && cargo clean",
     "cd ../scarlet-rt && cargo clean",
     "cd ../scarlet-os && cargo clean",
+    "cd ../scarlet-codecs && cargo clean",
     "cd ../std && cargo clean",
     "cd ../framebuffer && cargo clean",
     "cd ../gpu && cargo clean",
