+++ user/lib/scarlet-codecs/src/lib.rs
@@ -1,0 +1,35 @@
+//! Scarlet userspace codec request builders.
+//!
+//! This crate keeps codec parsing and DPB management in userspace. Kernel
+//! video backends receive prepared request parameters and only lower them to
+//! hardware-specific command streams.
+
+#![no_std]
+
+extern crate alloc;
+
+/// H.264/AVC stateless request construction.
+pub mod h264;
+
+pub use h264::{
+    H264PreparedAccessUnit, H264RequestContext, SCARLET_VIDEO_H264_DECODE_PARAM_FLAG_IDR,
+    SCARLET_VIDEO_H264_DPB_FLAG_LONG_TERM, SCARLET_VIDEO_H264_DPB_FLAG_VALID,
+    SCARLET_VIDEO_H264_PPS_FLAG_BOTTOM_FIELD_PIC_ORDER_IN_FRAME_PRESENT,
+    SCARLET_VIDEO_H264_PPS_FLAG_CONSTRAINED_INTRA_PRED,
+    SCARLET_VIDEO_H264_PPS_FLAG_DEBLOCKING_FILTER_CONTROL_PRESENT,
+    SCARLET_VIDEO_H264_PPS_FLAG_ENTROPY_CODING_MODE,
+    SCARLET_VIDEO_H264_PPS_FLAG_REDUNDANT_PIC_CNT_PRESENT,
+    SCARLET_VIDEO_H264_PPS_FLAG_TRANSFORM_8X8_MODE, SCARLET_VIDEO_H264_PPS_FLAG_WEIGHTED_PRED,
+    SCARLET_VIDEO_H264_SLICE_FLAG_DIRECT_SPATIAL_MV_PRED,
+    SCARLET_VIDEO_H264_SLICE_FLAG_REF_LISTS_PRESENT,
+    SCARLET_VIDEO_H264_SPS_FLAG_DELTA_PIC_ORDER_ALWAYS_ZERO,
+    SCARLET_VIDEO_H264_SPS_FLAG_DIRECT_8X8_INFERENCE, SCARLET_VIDEO_H264_SPS_FLAG_FRAME_CROPPING,
+    SCARLET_VIDEO_H264_SPS_FLAG_FRAME_MBS_ONLY,
+    SCARLET_VIDEO_H264_SPS_FLAG_GAPS_IN_FRAME_NUM_VALUE_ALLOWED,
+    SCARLET_VIDEO_H264_SPS_FLAG_MB_ADAPTIVE_FRAME_FIELD,
+    SCARLET_VIDEO_H264_SPS_FLAG_QPPRIME_Y_ZERO_TRANSFORM_BYPASS,
+    SCARLET_VIDEO_H264_SPS_FLAG_SEPARATE_COLOUR_PLANE, ScarletVideoH264DecodeParams,
+    ScarletVideoH264DpbEntry, ScarletVideoH264Pps, ScarletVideoH264PredWeights,
+    ScarletVideoH264Reference, ScarletVideoH264ScalingMatrix, ScarletVideoH264SliceParams,
+    ScarletVideoH264Sps, ScarletVideoH264StatelessParams, ScarletVideoH264WeightFactors,
+};
